refactor: extract table column editor
